+++ modules/UsingDynamicArray/ADTVector.c
@@ -64,7 +64,7 @@ void vector_set_at(Vector vec, int pos, Pointer value) {
 	assert(pos >= 0 && pos < vec->size);	// LCOV_EXCL_LINE (αγνοούμε το branch από τα coverage reports, είναι δύσκολο να τεστάρουμε το false γιατί θα κρασάρει το test)
 
 	// Αν υπάρχει συνάρτηση destroy_value, την καλούμε για το στοιχείο που αντικαθίσταται
-	if(vec->destroy_value != NULL)
+	if(value != vec->array[pos] && vec->destroy_value != NULL)
 		vec->destroy_value(vec->array[pos]);
 
 	vec->array[pos] = value;
